+++ fuel-tx/src/transaction/types/input/coin.rs
@@ -19,6 +19,8 @@ use fuel_types::{
     Word,
 };
 
+use super::PredicateCode;
+
 pub type CoinFull = Coin<Full>;
 pub type CoinSigned = Coin<Signed>;
 pub type CoinPredicate = Coin<Predicate>;
@@ -39,7 +41,7 @@ pub trait CoinSpecification: private::Seal {
         + for<'a> Compressible<
             Compressed: Clone + serde::Serialize + serde::Deserialize<'a>,
         >;
-    type Predicate: AsField<Vec<u8>>
+    type Predicate: AsField<PredicateCode>
         + Clone
         + for<'a> Compressible<
             Compressed: Clone + serde::Serialize + serde::Deserialize<'a>,
@@ -58,7 +60,7 @@ pub trait CoinSpecification: private::Seal {
 #[cfg(not(feature = "da-compression"))]
 pub trait CoinSpecification: private::Seal {
     type Witness: AsField<u16>;
-    type Predicate: AsField<Vec<u8>>;
+    type Predicate: AsField<PredicateCode>;
     type PredicateData: AsField<Vec<u8>>;
     type PredicateGasUsed: AsField<Word>;
 }
@@ -72,7 +74,7 @@ pub trait CoinSpecification: private::Seal {
 pub struct Signed;
 
 impl CoinSpecification for Signed {
-    type Predicate = Empty<Vec<u8>>;
+    type Predicate = Empty<PredicateCode>;
     type PredicateData = Empty<Vec<u8>>;
     type PredicateGasUsed = Empty<Word>;
     type Witness = u16;
@@ -87,7 +89,7 @@ impl CoinSpecification for Signed {
 pub struct Predicate;
 
 impl CoinSpecification for Predicate {
-    type Predicate = Vec<u8>;
+    type Predicate = PredicateCode;
     type PredicateData = Vec<u8>;
     type PredicateGasUsed = Word;
     type Witness = Empty<u16>;
@@ -98,7 +100,7 @@ impl CoinSpecification for Predicate {
 pub struct Full;
 
 impl CoinSpecification for Full {
-    type Predicate = Vec<u8>;
+    type Predicate = PredicateCode;
     type PredicateData = Vec<u8>;
     type PredicateGasUsed = Word;
     type Witness = u16;